+++ packages/build-info/src/browser/file-system.ts
@@ -1,4 +1,4 @@
-import { DirType, FileSystem } from '../file-system.js'
+import { DirType, Environment, FileSystem } from '../file-system.js'
 
 /** A sample implementation of a GitHub provider */
 export class GithubProvider {
@@ -38,6 +38,10 @@ export class GithubProvider {
 export class WebFS extends FileSystem {
   constructor(public git: GithubProvider) {
     super()
+  }
+
+  getEnvironment() {
+    return Environment.Browser
   }
 
   isAbsolute(path: string): boolean {