+++ packages/build-info/src/build-systems/buck.ts
@@ -1,5 +1,3 @@
-import { Project } from '../project.js'
-
 import { BaseBuildTool } from './build-system.js'
 
 export class Buck extends BaseBuildTool {
@@ -7,8 +5,11 @@ export class Buck extends BaseBuildTool {
   name = 'Buck'
   configFiles = ['.buckconfig', 'BUCK']
 
-  async detect(project: Project) {
-    const config = await project.fs.findUp(this.configFiles, { cwd: project.baseDirectory, stopAt: project.root })
+  async detect() {
+    const config = await this.project.fs.findUp(this.configFiles, {
+      cwd: this.project.baseDirectory,
+      stopAt: this.project.root,
+    })
 
     if (config) {
       return this