+++ packages/build-info/tests/failing-framework-info.test.ts
@@ -1,8 +1,15 @@
 import { afterEach, beforeEach, expect, test, vi } from 'vitest'
 
+import * as metrics from '../src/metrics.js'
 import { getBuildInfo } from '../src/node/get-build-info.js'
 
 import { createFixture } from './helpers.js'
+
+vi.mock('@netlify/framework-info', () => ({
+  listFrameworks: async () => {
+    throw new Error('Crashing hard 💣')
+  },
+}))
 
 const env = { ...process.env }
 beforeEach(() => {
@@ -13,32 +20,24 @@ beforeEach(() => {
 afterEach(async ({ cleanup }) => await cleanup?.())
 
 test('framework info should not crash build-info', async (ctx) => {
+  const metricsSpy = vi.spyOn(metrics, 'report').mockImplementation(vi.fn())
   const fixture = await createFixture('pnpm-simple')
   ctx.cleanup = fixture.cleanup
 
-  vi.mock('@netlify/framework-info', () => ({
-    listFrameworks: async () => {
-      throw new Error('Crashing hard 💣')
-    },
-  }))
-
-  const { frameworks, packageManager } = await getBuildInfo('', fixture.cwd)
+  const { frameworks, packageManager } = await getBuildInfo({ projectDir: '', rootDir: fixture.cwd })
   expect(packageManager?.name).toBe('pnpm')
   expect(frameworks).toEqual([])
+  expect(metricsSpy).toHaveBeenCalledWith(new Error('Crashing hard 💣'))
 })
 
 test('framework info should not crash build-info and the detection should still work', async (ctx) => {
+  const metricsSpy = vi.spyOn(metrics, 'report').mockImplementation(vi.fn())
   const fixture = await createFixture('yarn-project')
   ctx.cleanup = fixture.cleanup
 
-  vi.mock('@netlify/framework-info', () => ({
-    listFrameworks: async () => {
-      throw new Error('Crashing hard 💣')
-    },
-  }))
-
   process.env.NETLIFY_USE_PNPM = 'true'
-  const { frameworks, packageManager } = await getBuildInfo('', fixture.cwd)
+  const { frameworks, packageManager } = await getBuildInfo({ projectDir: '', rootDir: fixture.cwd })
   expect(packageManager?.name).toBe('pnpm')
   expect(frameworks).toEqual([])
+  expect(metricsSpy).toHaveBeenCalledWith(new Error('Crashing hard 💣'))
 })